Day10/EmployeeRequestTrackerSolution/RequestTrackerBLTest/DepartmentBLTest.cs
using RequestTrackerBLLibrary;
using RequestTrackerDALLibrary;
using RequestTrackerModelLibrary;

namespace RequestTrackerBLTest
{
    public class Tests
    {
        IRepository<int, Department> repository;
        IDepartmentService departmentService;
        [SetUp]
        public void Setup()
        {
            repository = new DepartmentRepository();
            
            departmentService = new DepartmentBL(repository);
        }

        [Test]
        public void GetDepartmnetByNameTest()
        {
            //Action
            var department = departmentService.GetDepartmentByName("IT");
            //Assert
            Assert.AreEqual(1, department.Id);
        }

        [Test]
        public void GetDepartmnetByNameExceptionTest()
        {
            //Action
            var exception = Assert.Throws<DepartmentNotFoundException>(() => departmentService.GetDepartmentByName("Admin"));
            //Assert
            Assert.AreEqual("The department could not be found", exception.Message);
        }
        [Test]
        public void AddDepartment()
        {
            //Action
            Department department = new Department() { Name = "IT", Department_Head = 101 };
            int id = departmentService.AddDepartment(department);
            //Assert
            Assert.AreEqual(1, id);
        }



    }
}